+++ src/features/auth/api/logout.ts
@@ -1,0 +1,8 @@
+import { API_END_POINT, httpClient, removeToken } from "@/shared";
+import { refreshToken } from ".";
+
+export const logout = async () => {
+  await refreshToken();
+  await httpClient.post(API_END_POINT.LOGOUT);
+  removeToken();
+};
