+++ src/features/user/ui/PreAuctionMyRegister.tsx
@@ -1,9 +1,9 @@
 import type { IPreAuctionRegisteredItem } from '@/@types/AuctionItem';
 import { LikeCount, Price, ROUTES } from '@/shared';
 import { useNavigate } from 'react-router-dom';
-import ProductItem from '../../entities/product/ui/ProductItem';
+import ProductItem from '../../../entities/product/ui/ProductItem';
 
-const PreAuctionMyRegister = ({ product }: { product: IPreAuctionRegisteredItem }) => {
+export const PreAuctionMyRegister = ({ product }: { product: IPreAuctionRegisteredItem }) => {
   const navigate = useNavigate()
   const clickProduct = () => navigate(ROUTES.PRE_AUCTION.getItemRoute(product.productId))
 
@@ -14,5 +14,3 @@ const PreAuctionMyRegister = ({ product }: { product: IPreAuctionRegisteredItem 
     </ProductItem>
   );
 };
-
-export default PreAuctionMyRegister;
